# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TESIS - Sistema de Clasificación de Riesgo",
        "description": "Documentación de la API del sistema de tesis",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/clasificaciones/{id}/pdf": {
            "get": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Generar reporte PDF de clasificación",
                "description": "Genera el reporte de clasificación en PDF y lo muestra directamente en el navegador",
                "operationId": "321c13d6d6616288f53fdb495dd2f93f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clasificación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF generado correctamente",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "404": {
                        "description": "Clasificación no encontrada"
                    }
                }
            }
        },
        "/api/clasificaciones/{id}/enviar-whatsapp": {
            "post": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Enviar enlace del PDF por WhatsApp",
                "description": "Envía por WhatsApp el enlace del reporte PDF generado para una clasificación",
                "operationId": "9e06a36c708586314256c9c3e4c5f52b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clasificación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telefono"
                                ],
                                "properties": {
                                    "telefono": {
                                        "description": "Número de teléfono con código de país",
                                        "type": "string",
                                        "example": "+51999999999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reporte enviado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al enviar el mensaje por WhatsApp"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingresa tu token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}